+++ src/main/java/com/kpei/mkbd/datatransform/dto/Vd57cDto.java
@@ -2,15 +2,17 @@ package com.kpei.mkbd.datatransform.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class Vd57cDto extends Vd5Base {
-    private double saldo;
-    private double lmHrKerja;
-    private double dimiliki;
-    private double dipisahkan;
+    private BigDecimal saldo;
+    private BigDecimal lmHrKerja;
+    private BigDecimal dimiliki;
+    private BigDecimal dipisahkan;
     private String penjelasan;
 }
